requirements.txt
Flask==2.3.2
python-docx
gunicorn
